Create master layout page
Because:
- It's easy to compose pages
- It appears more perfomant
- Enables DRY coding

I made a master layout component containing the header and footer.
Since both components permeate the project I figured it would be
better to create this component and re-use it when I need to
create any page, since it will become a question of simply
sandwiching the components I want to compose the page with in-
between the layout component.